up service impl

+++ src/main/java/com/example/palindromos/adapter/repository/PalavraRepository.java
@@ -3,5 +3,5 @@ package com.example.palindromos.adapter.repository;
 import com.example.palindromos.adapter.entities.PalavraEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PalavraRepository extends JpaRepository<PalavraEntity, Integer> {
+public interface PalavraRepository extends JpaRepository<PalavraEntity, Long> {
 }
